+++ ownership-chain/runtime/laos/src/configs/mod.rs
@@ -7,6 +7,7 @@ pub mod fee;
 mod multisig;
 pub mod parachain_staking;
 mod proxy;
+mod transaction_payment;
 
 // Define storage fees as constants for clarity and reuse
 const STORAGE_ITEM_FEE: Balance = 10 * UNIT;
